requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
psycopg2-binary==2.9.7
Pillow==10.0.1
requests==2.31.0
celery==5.3.1
redis==4.6.0
PyJWT==2.8.0
python-dotenv==1.0.0
prometheus-client==0.17.1
sqlparse==0.4.4
asgiref==3.7.2
tzdata==2023.3
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
pytest-mock==3.11.1
requests-mock==1.11.0
factory-boy==3.3.0
dj-database-url==2.0.0